refactor: Modularize application architecture
This commit refactors the entire application from a single monolithic `index.tsx` file into a modular and organized structure within a new `src` directory.

- All React components, views, hooks, and utility functions have been extracted into their own respective files under `src/`.
- TypeScript types and application configurations are now centralized in `src/types` and `src/utils`.
- The main `index.tsx` has been rewritten to be a clean entry point that assembles the application from its modular parts.
- Fixed several TypeScript errors and potential runtime bugs that were identified during the initial analysis.
